refactor: only specify one set schema
